% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_species.R
\name{mapping_species}
\alias{mapping_species}
\title{Mapping species}
\usage{
mapping_species(static_location, species)
}
\arguments{
\item{species}{}
}
\value{
an interactive map of species
}
\description{
Mapping species
}
\examples{
#leave blank for now

}
