% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_taxon.R
\name{mapping_taxon}
\alias{mapping_taxon}
\title{Mapping taxon}
\usage{
mapping_taxon(veg_df, taxon)
}
\arguments{
\item{taxon}{}
}
\value{
an interactive map of taxon
}
\description{
Mapping taxon
}
\examples{
#leave blank for now


TO DO: make a function that predicts the species for genus level obs.
}
