% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prism_annual_normals.R
\name{prism_annual_normals}
\alias{prism_annual_normals}
\title{PRISM annual normals for ecosite}
\usage{
prism_annual_normals(ecosite, prism_dir)
}
\arguments{
\item{ecosite}{ecosite id code}

\item{prism_dir}{directory to store PRISM data}
}
\value{
dataframe of summary stats
}
\description{
PRISM annual normals for ecosite
}
\examples{
ecositerSpatial::prism_normals(ecosite = "F022AB100CA", prism_dir = "C:/Users/Nathan.Roe/Documents/PRISM_R/")
}
