% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_prism_annual_normals.R
\name{site_prism_annual_normals}
\alias{site_prism_annual_normals}
\title{Extract PRISM annual normals for sites}
\usage{
site_prism_annual_normals(site_df, prism_dir, id, x, y, EPSG)
}
\arguments{
\item{site_df}{a dataframe containing geographic data - for ecositer users, most likely a veg_df which contains eastings/northings.}

\item{prism_dir}{location of PRISM data (either exists - data will not be downloaded, or non-existing - data will be downloaded)}

\item{x}{x coordinate (i.e., easting or longitude) column from site_df}

\item{y}{y coordinate (i.e., norhting or latitude) column from site_df}

\item{EPSG}{the EPSG code for the coordinate reference system (CRS) of the input data, in format EPSG:36211}
}
\value{
dataframe of site
}
\description{
This function allows you to extract PRISM normals for all the sites in your dataframe
}
\examples{
site_prism_annual_normals(prism_dir = "C:/Users/Nathan.Roe/Documents/PRISM_R/annual", id = siteiid, x = "utmeasting", y = "utmnorthing", EPSG = "EPSG:32611")
}
