% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_taxon.R
\name{mapping_taxon}
\alias{mapping_taxon}
\title{Mapping taxon}
\usage{
mapping_taxon(veg_df, taxon, x, y, EPSG)
}
\arguments{
\item{veg_df}{vegetation dataframe, likely created using \code{ecositer::create_veg_df()}}

\item{taxon}{taxon of interest}

\item{x}{x coordinate (i.e., easting or longitude) column from site_df}

\item{y}{y coordinate (i.e., northing or latitude) column from site_df}

\item{EPSG}{the EPSG code for the coordinate reference system (CRS) of the input data, in format EPSG:36211}
}
\value{
an interactive map of taxon
}
\description{
Mapping taxon
}
\examples{
ecositerSpatial::mapping_taxon(veg_df = my_veg_df, taxon = "Pinus", x = "utmeasting", y = "utmnorthing", EPSG = "EPSG:32611")

}
